Build Build/InstallDesktopFile at "make all" time
If you run stack as root (e.g. for "make install"), any files it
creates under ./ will, of course, be owned by root.  That's a
problem for subsequent runs as non-root.

Reduce the likelihood of that happening by building
Build/InstallDesktopFile during "make all", so that it needn't be
built by "make install".