remove no longer needed uuid check in prepSocket
Since 3dd43df9c2748df355314cd07b2bb8ebd30be69e, the socket warmup does
not run git-annex-shell on the remote host, and the point of this check
was to avoid error messages running git-annex-shell when it was not
installed. So the check is not needed any longer.

Also, this is one of only two uses of remoteGitConfig, which
I want to get rid of for reasons explained in
fc5888300fdbea52592bbb27c80f41673bee9d13.

This commit was sponsored by Fernando Jimenez on Patreon.