block url downloads by default
git-annex will refuse to download content from the web, to prevent
accidental exposure of data on private webservers on localhost and the
LAN. This can be overridden with the
annex.security.allowed-http-addresses setting.

This is the simplest possible fix for the security hole. A better fix
has been developed for newer versions of git-annex but would be a lot of
work to backport, and perhaps too big a diff.

There are several sets of git-annex users who will be impacted
in different ways by this:

* Users who have a git-annex repository but don't use the web special
  remote. Unaffected.

* Users who have a git-annex repository that is for private use only.
  They will have to read enough docs to find the setting to allow
  git annex addurl to work again.

* Users who have a git-annex repositry that is shared with people they
  don't fully trust.
  They will not be able to use the web special remote with this version
  of git-annex. They'll have to upgrade.

The S3, glacier, and webdav special remotes are still allowed to
download from the web. There are other potential attacks involving the
web server they connect to redirecting to a local private web server,
and tricking them from downloading content from it which then leaks back
to the attacker. Those attacks are not addressed here, but they also
seem fairly unlikely. Further analysis is needed; preliminary analysis
of glacier-cli, for example, suggests it does not follow redirects and
so is not vulnerable to such attacks.